Switch to wyhash from CRC for capture tile diff
Originally suggested by @trishume at #167.

(cherry picked from commit 813e613039bb451da2fe5916b87710b93d4e02da)

+++ module/rdpClientCon.h
@@ -123,7 +123,7 @@ struct _rdpClientCon
     RegionPtr dirtyRegion;
 
     int num_rfx_crcs_alloc[16];
-    int *rfx_crcs[16];
+    uint64_t *rfx_crcs[16];
     int send_key_frame[16];
 
     /* true = skip drawing */
